+++ src/Tribe/Views/V2/Hooks.php
@@ -23,7 +23,6 @@ use Tribe\Events\Views\V2\Repository\Event_Period;
 use Tribe\Events\Views\V2\Template\Featured_Title;
 use Tribe\Events\Views\V2\Template\Title;
 use Tribe\Events\Views\V2\Utils\View as View_Utils;
-use Tribe\Events\Views\V2\iCalendar\Subscribe as Subscribe;
 use Tribe__Context as Context;
 use Tribe__Customizer__Section as Customizer_Section;
 use Tribe__Events__Main as TEC;
@@ -68,7 +67,6 @@ class Hooks extends \tad_DI52_ServiceProvider {
 		add_action( 'the_post', [ $this, 'manage_sensitive_info' ] );
 		add_action( 'get_header', [ $this, 'print_single_json_ld' ] );
 		add_action( 'tribe_template_after_include:events/v2/components/after', [ $this, 'action_add_promo_banner' ], 10, 3 );
-		add_action( 'after_setup_theme', [ $this, 'after_setup_theme' ] );
 	}
 
 	/**
@@ -136,10 +134,6 @@ class Hooks extends \tad_DI52_ServiceProvider {
 
 		// iCalendar export request handling.
 		add_filter( 'tribe_ical_template_event_ids', [ $this, 'inject_ical_event_ids' ] );
-	}
-
-	public function after_setup_theme() {
-		tribe( Subscribe::class )->register();
 	}
 
 	/**
