+++ src/Tribe/Views/V2/Service_Provider.php
@@ -42,7 +42,7 @@ class Service_Provider extends \tad_DI52_ServiceProvider {
 
 		tribe_register_provider( Widgets\Service_Provider::class );
 		tribe_register_provider( Customizer\Service_Provider::class );
-		tribe_register_provider( iCalendar\Subscribe::class );
+		tribe_register_provider( iCalendar\Service_Provider::class );
 
 		$this->register_hooks();
 		$this->register_assets();
