+++ src/resources/postcss/components/full/_ical-link.pcss
@@ -108,7 +108,7 @@
 			.tec-subscribe-dropdown__selector-list-item:hover,
 			.tec-subscribe-dropdown__selector-list-item:focus,
 			.tec-subscribe-dropdown__selector-list-item:focus-within {
-				background-color: pink;/* var(--tec-color-accent-primary-background); */
+				background-color: var(--tec-color-background-secondary-datepicker);
 			}
 
 			.tec-subscribe-dropdown__selector-list-item-link:active,
